Change assertions prefix (#3)
This change aims to make migrations from Jest or Chai a lot less traumatic. Most of their assertions are written in a `toBe...` fashion and developers are already used to that prefix. If we stick to that standard, moving from those libraries will be easier and for some it'll be a big plus to have the look and feel from Jest